requirements.txt
begins
nltk
sklearn
wikipedia
